+++ src/Glpi/Api/HL/Controller/SetupController.php
@@ -37,6 +37,7 @@ namespace Glpi\Api\HL\Controller;
 
 use AuthLDAP;
 use CommonDBTM;
+use Config;
 use Glpi\Api\HL\Doc as Doc;
 use Glpi\Api\HL\Middleware\ResultFormatterMiddleware;
 use Glpi\Api\HL\ResourceAccessor;
@@ -51,6 +52,8 @@ final class SetupController extends AbstractController
 {
     public static function getRawKnownSchemas(): array
     {
+        global $DB;
+
         return [
             'LDAPDirectory' => [
                 'x-version-introduced' => '2.0',
@@ -104,6 +107,44 @@ final class SetupController extends AbstractController
                     ],
                 ],
             ],
+            'Config' => [
+                'x-version-introduced' => '2.1',
+                'x-itemtype' => Config::class,
+                'type' => Doc\Schema::TYPE_OBJECT,
+                'properties' => [
+                    'id' => [
+                        'type' => Doc\Schema::TYPE_INTEGER,
+                        'format' => Doc\Schema::FORMAT_INTEGER_INT64,
+                        'readOnly' => true,
+                    ],
+                    'context' => ['type' => Doc\Schema::TYPE_STRING],
+                    'name' => ['type' => Doc\Schema::TYPE_STRING],
+                    'value' => ['type' => Doc\Schema::TYPE_STRING],
+                ],
+                'x-rights-conditions' => [
+                    'read' => static function () use ($DB) {
+                        // Make a SQL request to get all config items so we can check which are undisclosed
+                        // We are using safe IDs rather than undisclosed IDs to avoid issues with concurrent modifications
+                        // We cannot reliably lock the table due to the fact that the DB connection here may differ from the one used to perform the actual read in the Search code
+                        $disclosed_ids = [];
+
+                        $it = $DB->request([
+                            'SELECT' => ['id', 'context', 'name'],
+                            'FROM'   => 'glpi_configs',
+                        ]);
+                        $test_configs = [];
+                        foreach ($it as $row) {
+                            $test_configs[] = $row + ['value' => 'dummy'];
+                        }
+                        foreach ($test_configs as $f) {
+                            if (!self::isUndisclosedConfig($f['context'], $f['name'])) {
+                                $disclosed_ids[] = $f['id'];
+                            }
+                        }
+                        return ['WHERE' => ['_.id' => $disclosed_ids]];
+                    },
+                ],
+            ],
         ];
     }
 
@@ -118,6 +159,7 @@ final class SetupController extends AbstractController
         if ($types === null) {
             $types = [
                 'LDAPDirectory' => AuthLDAP::getTypeName(1),
+                // Do not add Config here as it is handled specially
             ];
         }
         return $types_only ? array_keys($types) : $types;
@@ -211,5 +253,108 @@ final class SetupController extends AbstractController
     {
         $itemtype = $request->getAttribute('itemtype');
         return ResourceAccessor::deleteBySchema($this->getKnownSchema($itemtype, $this->getAPIVersion($request)), $request->getAttributes(), $request->getParameters());
+    }
+
+    private static function isUndisclosedConfig(string $context, string $name): bool
+    {
+        $f = ['context' => $context, 'name' => $name, 'value' => 'dummy'];
+        Config::unsetUndisclosedFields($f);
+        return !array_key_exists('value', $f);
+    }
+
+    #[Route(path: '/Config/{context}/{name}', methods: ['PATCH'], requirements: [
+        'context' => '\w+',
+        'name' => '\w+',
+    ], middlewares: [ResultFormatterMiddleware::class])]
+    #[RouteVersion(introduced: '2.1')]
+    #[Doc\UpdateRoute(schema_name: 'Config')]
+    public function setConfigValue(Request $request): Response
+    {
+        // Skip using ResourceAccessor given the particularities of Config
+        if (!Config::canUpdate()) {
+            return AbstractController::getAccessDeniedErrorResponse();
+        }
+        $context = $request->getAttribute('context');
+        $name = $request->getAttribute('name');
+        $value = $request->getParameter('value');
+        Config::setConfigurationValues($context, [$name => $value]);
+        // Return the updated config
+        if (self::isUndisclosedConfig($context, $name)) {
+            // If the field is undisclosed, only return a 204 to indicate success without revealing the value
+            return new JSONResponse(null, 204);
+        }
+        return new JSONResponse([
+            'context' => $context,
+            'name'    => $name,
+            'value'   => Config::getConfigurationValue($context, $name),
+        ]);
+    }
+
+    #[Route(path: '/Config', methods: ['GET'], middlewares: [ResultFormatterMiddleware::class])]
+    #[RouteVersion(introduced: '2.1')]
+    #[Doc\SearchRoute(schema_name: 'Config')]
+    public function searchConfigValues(Request $request): Response
+    {
+        return ResourceAccessor::searchBySchema($this->getKnownSchema('Config', $this->getAPIVersion($request)), $request->getParameters());
+    }
+
+    #[Route(path: '/Config/{context}', methods: ['GET'], requirements: [
+        'context' => '\w+',
+    ], middlewares: [ResultFormatterMiddleware::class])]
+    #[RouteVersion(introduced: '2.1')]
+    #[Doc\SearchRoute(schema_name: 'Config')]
+    public function searchConfigValuesByContext(Request $request): Response
+    {
+        $filters = $request->hasParameter('filter') ? $request->getParameter('filter') : '';
+        $filters .= ';context==' . $request->getAttribute('context');
+        $request->setParameter('filter', $filters);
+        return ResourceAccessor::searchBySchema($this->getKnownSchema('Config', $this->getAPIVersion($request)), $request->getParameters());
+    }
+
+    #[Route(path: '/Config/{context}/{name}', methods: ['GET'], requirements: [
+        'context' => '\w+',
+        'name' => '\w+',
+    ], middlewares: [ResultFormatterMiddleware::class])]
+    #[RouteVersion(introduced: '2.1')]
+    #[Doc\GetRoute(schema_name: 'Config')]
+    public function getConfigValue(Request $request): Response
+    {
+        // Skip using ResourceAccessor given the particularities of Config
+        $context = $request->getAttribute('context');
+        $name = $request->getAttribute('name');
+        $config = new Config();
+        if (!$config->getFromDBByCrit(['context' => $context, 'name'    => $name,])) {
+            return AbstractController::getNotFoundErrorResponse();
+        }
+        if (self::isUndisclosedConfig($context, $name) || !$config->can($config->getID(), READ)) {
+            return AbstractController::getAccessDeniedErrorResponse();
+        }
+        return new JSONResponse([
+            'context' => $context,
+            'name'    => $name,
+            'value'   => Config::getConfigurationValue($context, $name),
+        ]);
+    }
+
+    #[Route(path: '/Config/{context}/{name}', methods: ['DELETE'], requirements: [
+        'context' => '\w+',
+        'name' => '\w+',
+    ])]
+    #[RouteVersion(introduced: '2.1')]
+    #[Doc\DeleteRoute(schema_name: 'Config')]
+    public function deleteConfigValue(Request $request): Response
+    {
+        // Skip using ResourceAccessor given the particularities of Config
+        if (!Config::canUpdate()) {
+            return AbstractController::getAccessDeniedErrorResponse();
+        }
+        $context = $request->getAttribute('context');
+        $name = $request->getAttribute('name');
+        $config = new Config();
+        if (!$config->getFromDBByCrit(['context' => $context, 'name' => $name])) {
+            return AbstractController::getNotFoundErrorResponse();
+        }
+        Config::deleteConfigurationValues($context, [$name]);
+        return new JSONResponse(null, 204);
     }
 }